-Moved http request to http service -Added temporary input validation to disable `Extract` button -Fixed http request to ceneo
